python_codes/requirements.txt
torch
transformers
datasets
evaluate
accelerate
timm
pyyaml